PyDriller pipeline added
PyDriller part of the pipeline is added. PyDriller is used for cloning
the project then mines the commit of the repository based on a given
date interval. Among the commits only "*.java" file modifications are
downloaded. The downloaded files are organized by commit hash value and
"before/after" change folders. These files will be parsed by the change
inspector to identify the change types.