Delete Abstract Modifier (DAbM) method rule added
If an "abstract" modifier is removed from a method in a change rule
added. Also the Add Method (AM) rule strategy is changed. The previous
rule checked if the numbers of method has increase. However, in a
scenario where many methods are deleted and one new method is added
would failed. Therefore, the rule is changed in a way by deleting all
the before change methods from after change methods list. If still
methods remain in after change that means new methods are added.