Change Return Type (CRM) method rule added
Whenever the return type of the method is changed, and the method name,
parameter types are not changed, it will be classified under this rule.
Additional attributes added to into MethodInformation, parameter name
list and parameter type list. Parameter names might change but this will
not affect the method signature. Therefore these attributes are added,
which also affected the general comparison. Currently, whenever a CRM is
detected an AM is detected as well. In other words, CRM is a subset of
AM. However, if we want to decouple these two change types from each
other, the AM strategy must be changed again.