Change Name of Parameter Method (CNPM) rule added
When a method's only parameters name(s) is/are changed, it will be
detected by this rule. It assumes that method return type, method name,
method parameter types are not changed. Basically the method signature
does not change only the the parameter name(s) changes. In addition, the
parameter collection in MethodClassDataCollector issue has been fixed.
It wasn't able to detect the number of parameters above 2.