Change Parameter Method (CPM) rule added
This is a change of the type of the parameter, sequence of the parameter
and number of the parameter. The basis is the method name, method return
does not change, but also the before change method is no longer
preserved. Therefore, we remove all the existing method from before
change. Also, the first condition checks even if the bodies are same or
not.