Change Return Type Method (CRM) bug fixed
CRM has looking for containing all parameters types, which means the
compared method could be a super set of the other method in terms of
parameters. However, the parameter types need to be exactly same.
Therefore, it classified wrong some changes wrong. The containsAll
comparing strategy is changed to equals.