If-Statement class definition revised
Expression is used as the conditions of the if-statements, and could
exist of multiple expressions. On the other hand, the Statements is
consisted of list of expressions and used for the body of if-blocks.
So, the expression list in the IfStatement class is the list of
condition expressions, while the list of statements are the body of the
if-statement. Also equals methods added for if-statement, one for entire
if statement, one for only checking the conditions, and one fore
checking the body of if-statement.